React exercises
All can be run if index.js render option and import are changed to name of the task.  On default there is Sliders.js.
Binary.js changes the binary number by checking the boxes.
Movie.js and Movie.css shows movies on page, that can be filtered with category. NewCanvas draws lines when pressing a button. Sliders changes the background color based on the slider values that are rgb values.